Add Node class for representing nodes in the config hierarchy
This class is intended to have dynamic (but strong) typing and be explicit when modifying values. The attach/detach methods, in particular, exits to make the necessary behavior of shifting array elements in order to avoid an unserializable state. The root is not attached by default to avoid having to change the type when attaching children (as attaching a node will create the necessary structure).