Add base serialization system (Serializer and Deserializer)
Deserializer exists as a separate interface for serializers like optional and union (which are narrowing conversions) to require developers to explicitly specify the behavior before being used as a serializer. The Serializer interface extends Deserializer as all serializable values should be deserializable.

Tests for these will be incorporated with the serializers for primitive values (integer, string, etc) in a future commit.