Add nullable and optional serializers
These serializers are tricky because they have to account for both serialization and deserialization (something missed with collections) as well as requiring explicit reserialization behavior. The case of nullable and optional combined has not entirely been addressed either. Helper methods should be added to the Deserializer/Serializer interfaces as well.