Change default behavior of attach/detach to 'if necessary' (Resolves #1)
The majority of use cases for attach/detach used the 'if necessary' behavior. The current state of the Node generally shouldn't be known (or needed) for serialization.

Attach may still fail when overwriting, but this can only happen if the incoming node is at an invalid location. This should only happen when the Node class is being used directly, as serializers ensure that any child node accessed has the proper parent type.