Add tests for Storm parser/generator, fixing the following issues:
 - Lexer:
    - Missing Preconditions for invariants
    - Incorrect regex for identifiers
    - lexCharacter was just plain wrong (lexString copy)
    - lexString peek incorrectly excluded '\'
     - Missed the '\f' escape character
 - Parser:
    - Unicode escapes (`\uXXXX`) were not implemented
    - Leading newlines in the input
    - EOF directly after last value
 - Generator:
    - Repeated escape of '\' during generation
 - Not accounting for styling issues, misleading error messages, and other non-semantic issues.

This is not my first lexer. It is not my first parser. Nor is it my first for working with either JSON or Storm. This was manually tested on a few different inputs before being committed as well. Yet still, none of these issues were caught.

This is why you write tests.