Add detailed error diagnostics for the lexer/parser
This makes two important changes to how the parser and lexer work. First, information about the location of tokens (index + line + column) is now included and managed by the lexer. Second, a context is maintained which highlights important parts of the source that should be included in diagnostics (such as parent nodes). In the future, the parser and lexer should be communicating to share this contextual information (at the moment, the lexer tracks brackets and braces).

The parser currently does not do anything with error recovery, as fail-fast is sufficient for now. While there are some checks to provide better error messages in certain situations (particularly when hitting the end of a line), a method of testing these scenarios needs to be devised first.