Remove Deserializer interface (Resolves #3)
As outlined in the issue, this interface was well-intentioned but resulted in unforeseen api and implementation problems that made the library too complex.

While the static type safety is valuable, I believe it is not worth the resulting complexity. Reserialization should, in practice, only fail as a result of developer error, unlike deserialization where the input is controlled by the end user. Reserialization is also planned to include dynamic behavior already with class-based serialization, and serializers for objects/tuples will need to handle dynamic serialization already.

As part of this, the default reserialization behavior for Nullable and Optional serializers is to reserialize the defaults (so a value will be serialized as that value, not null/undefined). Having worked with the lexer/parser/generator, there are already a number of defaults being made or other things which cannot be represented the same on reserialization (unicode escapes for instance), so at the moment I find this acceptable.

As a final note, the is room to discuss removing SerializationException from the reserialize method entirely, since it is almost always a developer issue. I can think of metaprogramming systems as a potential exception here, so some additional thought should be given to it.