Add test cases for diagnostic range (#4)
This tests the range included in the diagnostic matches the expected range, and thus does not check error messages or causes.

For some errors, such as a missing separators, the range returned is the next token (which is expected to be a separator, but is instead the start of another value/property). In effect, this is testing *invalid* separators, not *missing*. In the future the parser should attempt to recover by trying to parse a value/property, and consider the separator as missing if at least one token is successfully consumed. The range returned here should be the end of the previous value, rather than the start of the new value (or perhaps both?).