Add Tuple and Object serializers for non-homogeneous collections
This makes use of dynamic serialization behavior since the exact types cannot be verified at compile time. Notably, this catches any ClassCastException thrown when delegating to a serializer - there is likely a better solution that can be found (especially once getting into class-based serialization), but for now this is acceptable.