Refactor StormParser to use nodes directly
Instead of deserializing a value and setting it to a Node at the end, this changes the parser to build up the Node during parsing. This is the preferred solution in general but is also necessary for setting comments later on.

Additionally, this adds a check for duplicate keys at the parser level, which is necessary to avoid overwriting data and include the context. A test case for this has been added as well.