Add node comments to the node hierarchy and serializers (#2 part 1)
Collection types (ARRAY/OBJECT) should maintain comments in their children, which is done by always passing the same node rather than clearing it. Children which are unused must be detached first, which must be done while avoiding a ConcurrentModificationException. Tests have been added for this behavior as well.

Implementation of comments in the parser is a WIP.