Fix/improve error ranges for whitespace-only input
Input that contains only whitespace is invalid in JSON (while valid in
Storm; being an empty object). The calculations of error ranges in these
cases has been fixed to avoid an IndexOutOfBoundsException when checking
tokens.has(-1). For empty input (or EOL), the diagnostic message will
include at least one '^' marker to show where the character should be.

In retrospect, there's a lot of cases where the ideal error range
depends on the specific error, such as missing commas (the best location
is probably the next character rather than next token).