Fix (partially) field order with class-based constructor serialization
Constructor serialization depends on parameters matching the order of
declared fields. This fixes fields being stored in a HashMap instead of
a LinkedHashMap, effectively loosing their order.

However, Class#getFields is not documented to return fields in a defined
order - rather, the order is implementation-dependent. In practice this
is often the order declared, but this is not guaranteed. A long-term
solution may consider using parameter names to determine order (which
also allows multiple constructors), however parameter names are only
available if compiled with the -parameters option.